+++ lib/db.bash
@@ -48,7 +48,7 @@ function odoo_db_create {
             ;;
             -h|--help|help)
                 echo "$usage";
-                exit 0;
+                return 0;
             ;;
             *)
                 break;
@@ -66,8 +66,8 @@ function odoo_db_create {
 
     echov -e "${BLUEC}Creating odoo database ${YELLOWC}$db_name${BLUEC} using conf file ${YELLOWC}$conf_file${NC}";
 
-    local python_cmd="import lodoo; cl=lodoo.Client(['-c', '$conf_file']);";
-    python_cmd="$python_cmd cl.db.create_database(cl._server.tools.config['admin_passwd'], '$db_name', $demo_data, '$db_lang');"
+    local python_cmd="import lodoo; cl=lodoo.LocalClient(['-c', '$conf_file']);";
+    python_cmd="$python_cmd cl.db.create_database(cl.odoo.tools.config['admin_passwd'], '$db_name', $demo_data, '$db_lang');"
 
     run_python_cmd "$python_cmd";
     
@@ -79,19 +79,24 @@ function odoo_db_drop {
     local db_name=$1;
     local conf_file=${2:-$ODOO_CONF_FILE};
 
-    local python_cmd="import lodoo; cl=lodoo.Client(['-c', '$conf_file']);";
-    python_cmd="$python_cmd cl.db.drop(cl._server.tools.config['admin_passwd'], '$db_name');"
+    if ! odoo_db_exists -q $db_name; then
+        echoe -e "${REDC}ERROR${NC}: Cannot drop database ${YELLOWC}${db_name}${NC}! Database does not exists!";
+        return 1;
+    fi
+
+    local python_cmd="import lodoo; cl=lodoo.LocalClient(['-c', '$conf_file']);";
+    python_cmd="$python_cmd cl.db.drop(cl.odoo.tools.config['admin_passwd'], '$db_name');"
     
     run_python_cmd "$python_cmd";
     
-    echoe -e "${GREENC}OK${NC}: Database ${BLUEC}$db_name${NC} dropt successfuly!";
+    echoe -e "${GREENC}OK${NC}: Database ${YELLOWC}$db_name${NC} dropt successfuly!";
 }
 
 # odoo_db_list [odoo_conf_file]
 function odoo_db_list {
     local conf_file=${1:-$ODOO_CONF_FILE};
 
-    local python_cmd="import lodoo; cl=lodoo.Client(['-c', '$conf_file', '--logfile', '/dev/null']);";
+    local python_cmd="import lodoo; cl=lodoo.LocalClient(['-c', '$conf_file', '--logfile', '/dev/null']);";
     python_cmd="$python_cmd print('\n'.join(['%s'%d for d in cl.db.list()]));";
     
     run_python_cmd "$python_cmd";
@@ -138,7 +143,7 @@ function odoo_db_exists {
     local db_name=$1;
     local conf_file=${2:-$ODOO_CONF_FILE};
 
-    local python_cmd="import lodoo; cl=lodoo.Client(['-c', '$conf_file', '--logfile', '/dev/null']);";
+    local python_cmd="import lodoo; cl=lodoo.LocalClient(['-c', '$conf_file', '--logfile', '/dev/null']);";
     python_cmd="$python_cmd exit(int(not(cl.db.db_exist('$db_name'))));";
     
     if run_python_cmd "$python_cmd"; then
@@ -156,8 +161,8 @@ function odoo_db_rename {
     local new_db_name=$2;
     local conf_file=${3:-$ODOO_CONF_FILE};
 
-    local python_cmd="import lodoo; cl=lodoo.Client(['-c', '$conf_file']);";
-    python_cmd="$python_cmd cl.db.rename(cl._server.tools.config['admin_passwd'], '$old_db_name', '$new_db_name');"
+    local python_cmd="import lodoo; cl=lodoo.LocalClient(['-c', '$conf_file']);";
+    python_cmd="$python_cmd cl.db.rename(cl.odoo.tools.config['admin_passwd'], '$old_db_name', '$new_db_name');"
     
     if run_python_cmd "$python_cmd"; then
         echoe -e "${GREENC}OK${NC}: Database ${BLUEC}$old_db_name${NC} renamed to ${BLUEC}$new_db_name${NC} successfuly!";
@@ -185,8 +190,8 @@ function odoo_db_dump {
         fi
     fi
 
-    local python_cmd="import lodoo; cl=lodoo.Client(['-c', '$conf_file']);";
-    python_cmd="$python_cmd dump=cl.db.dump(cl._server.tools.config['admin_passwd'], '$db_name' $format_opt).decode('base64');";
+    local python_cmd="import lodoo, base64; cl=lodoo.LocalClient(['-c', '$conf_file']);";
+    python_cmd="$python_cmd dump=base64.b64decode(cl.db.dump(cl.odoo.tools.config['admin_passwd'], '$db_name' $format_opt));";
     python_cmd="$python_cmd open('$db_dump_file', 'wb').write(dump);";
     
     if run_python_cmd "$python_cmd"; then
@@ -253,8 +258,8 @@ function odoo_db_restore {
     local db_dump_file=$2;
     local conf_file=${3:-$ODOO_CONF_FILE};
 
-    local python_cmd="import lodoo; cl=lodoo.Client(['-c', '$conf_file']);";
-    python_cmd="$python_cmd res=cl.db.restore(cl._server.tools.config['admin_passwd'], '$db_name', open('$db_dump_file', 'rb').read().encode('base64'));";
+    local python_cmd="import lodoo, base64; cl=lodoo.LocalClient(['-c', '$conf_file']);";
+    python_cmd="$python_cmd res=cl.db.restore(cl.odoo.tools.config['admin_passwd'], '$db_name', base64.b64encode(open('$db_dump_file', 'rb').read()));";
     python_cmd="$python_cmd exit(0 if res else 1);";
     
     if run_python_cmd "$python_cmd"; then
@@ -285,7 +290,7 @@ function odoo_db_command {
 
     if [[ $# -lt 1 ]]; then
         echo "$usage";
-        exit 0;
+        return 0;
     fi
 
     while [[ $# -gt 0 ]]
@@ -295,55 +300,55 @@ function odoo_db_command {
             list)
                 shift;
                 odoo_db_list "$@";
-                exit;
+                return;
             ;;
             create)
                 shift;
                 odoo_db_create "$@";
-                exit;
+                return;
             ;;
             drop)
                 shift;
                 odoo_db_drop "$@";
-                exit;
+                return;
             ;;
             dump)
                 shift;
                 odoo_db_dump "$@";
-                exit;
+                return;
             ;;
             backup)
                 shift;
                 odoo_db_backup "$@";
-                exit;
+                return;
             ;;
             backup-all)
                 shift;
                 odoo_db_backup_all "$@";
-                exit;
+                return;
             ;;
             restore)
                 shift;
                 odoo_db_restore "$@";
-                exit;
+                return;
             ;;
             exists)
                 shift;
                 odoo_db_exists "$@";
-                exit;
+                return;
             ;;
             rename)
                 shift;
                 odoo_db_rename "$@";
-                exit;
+                return;
             ;;
             -h|--help|help)
                 echo "$usage";
-                exit 0;
+                return 0;
             ;;
             *)
                 echo "Unknown option / command $key";
-                exit 1;
+                return 1;
             ;;
         esac
         shift