+++ lib/pylib/lodoo.py
@@ -3,10 +3,8 @@
 """
 import os
 import atexit
+import functools
 import pkg_resources
-
-
-import erppeek
 
 
 def get_odoo_pkg():
@@ -70,13 +68,6 @@ def start_odoo_services(options=None, appname='odoo-helper'):
     return odoo
 
 
-# Monkeypatch erppeeks start services to be compatible with odoo 10.0
-erppeek.start_odoo_services = start_odoo_services
-
-# look same as erppeek
-from erppeek import *
-
-
 class LocalModel(object):
     def __init__(self, client, name):
         self._client = client
@@ -88,25 +79,26 @@ class LocalModel(object):
         return wrapper
 
 
-class LocalClient(object):
-    """ Wrapper for local odoo instance
-    """
-    def __init__(self, db, options=None):
-        if options is None:
-            options = []
+class LocalRegistry(object):
+    def __init__(self, client, dbname):
+        self._client = client
+        self._dbname = dbname
 
-        self.odoo = start_odoo_services(options)
         if self.odoo._api_v7:
-            self.registry = self.odoo.modules.registry.RegistryManager.get(db)
+            self.registry = self.odoo.modules.registry.RegistryManager.get(self._dbname)
             self.cursor = self.registry.db.cursor()
             self._env = None
         else:
             # For odoo 8, 9, 10, +(?) there is special function `odoo.registry`
             # to get registry instance for db
-            self.registry = self.odoo.registry(db)
+            self.registry = self.odoo.registry(self._dbname)
             self.cursor = self.registry.cursor()
             self._env = self.odoo.api.Environment(
                 self.cursor, self.odoo.SUPERUSER_ID, {})
+
+    @property
+    def odoo(self):
+        return self._client.odoo
 
     @property
     def env(self):
@@ -140,6 +132,90 @@ class LocalClient(object):
         self.env[model]._parent_store_compute()
         self.env.cr.commit()
 
+    def compute_translation_rate(self, lang, addons):
+        trans = self.env['ir.translation'].search([
+            ('module', 'in', addons),
+            ('lang', '=', lang),
+        ])
+        bad_translations = trans.filtered(
+            lambda r: not r.value or
+                      not r.value.strip() or
+                      r.src == r.value or
+                      r.source == r.value)
+
+        rate_by_addon = {}
+        for addon in addons:
+            addon_data = rate_by_addon[addon] = {}
+
+            addon_data['terms_total'] = trans_total = len(trans.filtered(
+                lambda r: r.module == addon))
+            addon_data['terms_untranslated'] = trans_fail = len(
+                bad_translations.filtered(lambda r: r.module == addon))
+
+            if trans_total:
+                addon_data['rate'] = 1.0 - (float(trans_fail) /
+                                              float(trans_total))
+            else:
+                addon_data['rate'] = 0.0
+
+            addon_data['rate'] *= 100.0
+
+        if trans:
+            total_rate = 1.0 - float(len(bad_translations)) / float(len(trans))
+        else:
+            total_rate = 0.0
+
+        total_rate *= 100.0
+        return {
+            'total_rate': total_rate,
+            'terms_total': len(trans),
+            'terms_untranslated': len(bad_translations),
+            'by_addon': rate_by_addon,
+        }
+
+    def print_translation_rate(self, translation_rate):
+        """ Print translation rate computed by `compute_translation_rate`
+        """
+        name_col_width = max([len(i) for i in translation_rate['by_addon']])
+
+        header_format_str = "%%-%ds | %%10s | %%15s | %%+10s" % name_col_width
+        row_format_str = "%%-%ds | %%10s | %%15s | %%7.2f" % name_col_width
+        spacer_str = "-" * (name_col_width + 3 + 10 + 3 + 15 + 3 + 10)
+
+        # Print header
+        print (header_format_str % (
+               'Addon', 'Total', 'Untranslated', 'Rate'))
+        print (spacer_str)
+
+        # Print translation rate by addon
+        for addon, rate_data in translation_rate['by_addon'].items():
+            print(row_format_str % (
+                  addon, rate_data['terms_total'],
+                  rate_data['terms_untranslated'],
+                  rate_data['rate']))
+
+        # Print total translation rate
+        print (spacer_str)
+        print(row_format_str % (
+              'TOTAL', translation_rate['terms_total'],
+              translation_rate['terms_untranslated'],
+              translation_rate['total_rate'],
+        ))
+
+    def assert_translation_rate(self, rate, min_total_rate=None,
+                                min_addon_rate=None):
+        """ Check translation rate, and return number, that can be used as exit
+            code
+        """
+        if min_total_rate is not None and rate['total_rate'] < min_total_rate:
+            return 1;
+
+        if min_addon_rate is not None:
+            for addon, rate_data in rate['by_addon'].items():
+                if rate_data['rate'] < min_addon_rate:
+                    return 2;
+        return 0;
+
     def call_method(self, model, method, *args, **kwargs):
         """ Simple wrapper to call local model methods for database
         """
@@ -153,3 +229,62 @@ class LocalClient(object):
 
     def __getitem__(self, name):
         return LocalModel(self, name)
+
+
+class LocalDBService(object):
+    def __init__(self, client):
+        self._client = client
+        self._dispatch = None
+
+    @property
+    def odoo(self):
+        return self._client.odoo
+
+    @property
+    def dispatch(self):
+        if self._dispatch is None:
+            if self.odoo._api_v7:
+                self._dispatch = self.odoo.netsvc.ExportService.getService('db').dispatch
+            else:   # Odoo v8
+                self._dispatch = functools.partial(self.odoo.http.dispatch_rpc, 'db')
+        return self._dispatch
+
+    def __getattr__(self, name):
+        def db_service_method(*args):
+            return self.dispatch(name, args)
+        return db_service_method
+
+
+class LocalClient(object):
+    """ Wrapper for local odoo instance
+    """
+    def __init__(self, options=None):
+        if options is None:
+            options = []
+
+        self._options = options
+
+        self._odoo = None
+        self._registries = {}
+        self._db_service = None
+
+    @property
+    def odoo(self):
+        if self._odoo is None:
+            self._odoo = start_odoo_services(self._options)
+        return self._odoo
+
+    @property
+    def db(self):
+        if self._db_service is None:
+            self._db_service = LocalDBService(self)
+        return self._db_service
+
+    def get_registry(self, dbname):
+        registry = self._registries.get(dbname, None)
+        if registry is None:
+            registry = self._registries[dbname] = LocalRegistry(self, dbname)
+        return registry
+
+    def __getitem__(self, name):
+        return self.get_registry(name)